Add support for parsing lists with settings
