+++ Nemesis.TextParsers.Tests/LinqTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using System.Text;
+using Nemesis.TextParsers.Parsers;
 using Nemesis.TextParsers.Utils;
 using NUnit.Framework;
 
@@ -9,16 +10,20 @@ namespace Nemesis.TextParsers.Tests
     [TestFixture]
     class LinqTests
     {
-        private static ParsedSequence<double> GetSequence(string text) => 
-            SpanCollectionSerializer.DefaultInstance.ParseStream<double>(text.AsSpan());
-
+        private static readonly ITransformer<double> _doubleTransformer = DoubleParser.Instance;
+        private static ParsedSequence GetSequence(string text)
+        {
+            var tokens = text.AsSpan().Tokenize('|', '\\', true);
+            return tokens.Parse('\\', '∅', '|');
+        }
+        
         [TestCase(@"", false, 0)]
         [TestCase(@"1", true, 1)]
         [TestCase(@"1|-1|2|-2", true, 0)]
         [TestCase(@"1|2|3|4|5|6|7|8|9", true, 45)]
         public void Sum(string text, bool success, double result)
         {
-            var pair = GetSequence(text).Sum();
+            var pair = GetSequence(text).Sum(_doubleTransformer);
             Assert.That(pair.success, Is.EqualTo(success));
             Assert.That(pair.result, Is.EqualTo(result));
         }
@@ -30,7 +35,7 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"1|2|3|4|5|6|7|8|9", true, 5)]
         public void Average(string text, bool expectedSuccess, double expectedResult)
         {
-            var (success, result) = GetSequence(text).Average();
+            var (success, result) = GetSequence(text).Average(_doubleTransformer);
             Assert.That(success, Is.EqualTo(expectedSuccess));
             Assert.That(result, Is.EqualTo(expectedResult));
         }
@@ -45,7 +50,7 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"1|2|3|4|5|6|7|8|9|-1|-2|-3|-4|-5|-6|-7|-8|-9", true, 33.529411764705891d)]
         public void Variance(string text, bool success, double result)
         {
-            var pair = GetSequence(text).Variance();
+            var pair = GetSequence(text).Variance(_doubleTransformer);
             Assert.That(pair.success, Is.EqualTo(success));
             Assert.That(pair.result, Is.EqualTo(result).Within(2).Ulps);
         }
@@ -66,7 +71,7 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"NaN|NaN|NaN", true, double.NaN)]
         public void Max(string text, bool success, double result)
         {
-            var pair = GetSequence(text).Max();
+            var pair = GetSequence(text).Max(_doubleTransformer);
             Assert.That(pair.success, Is.EqualTo(success));
             Assert.That(pair.result, Is.EqualTo(result).Within(2).Ulps);
         }
@@ -87,7 +92,7 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"NaN|NaN|NaN", true, double.NaN)]
         public void Min(string text, bool success, double result)
         {
-            var pair = GetSequence(text).Min();
+            var pair = GetSequence(text).Min(_doubleTransformer);
             Assert.That(pair.success, Is.EqualTo(success));
             Assert.That(pair.result, Is.EqualTo(result).Within(2).Ulps);
         }
@@ -96,7 +101,7 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"1|2|3|4|5|6|7|8|9", true, 45)]
         public void Aggregate(string text, bool success, double result)
         {
-            var pair = GetSequence(text).Aggregate((a, b) => a + b);
+            var pair = GetSequence(text).Aggregate(_doubleTransformer, (a, b) => a + b);
             Assert.That(pair.success, Is.EqualTo(success));
             Assert.That(pair.result, Is.EqualTo(result).Within(2).Ulps);
         }
@@ -105,8 +110,8 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"1|2|3|4|5|6|7|8|9", "102030405060708090")]
         public void AggregateSeed(string text, string result)
         {
-            var actual = GetSequence(text).Aggregate(
-                new StringBuilder(), 
+            var actual = GetSequence(text).Aggregate(_doubleTransformer,
+                new StringBuilder(),
                 (sb, current) => sb.Append((current * 10.0).ToString(null, CultureInfo.InvariantCulture))
                 );
             Assert.That(actual.ToString(), Is.EqualTo(result));
@@ -116,10 +121,10 @@ namespace Nemesis.TextParsers.Tests
         [TestCase(@"1|2|3|4|5|6|7|8|9", "102030405060708090")]
         public void AggregateSeedResult(string text, string result)
         {
-            var actual = GetSequence(text).Aggregate(
-                new StringBuilder(), 
+            var actual = GetSequence(text).Aggregate(_doubleTransformer,
+                new StringBuilder(),
                 (sb, current) => sb.Append((current * 10.0).ToString(null, CultureInfo.InvariantCulture)),
-                sb=>sb.ToString()
+                sb => sb.ToString()
                 );
             Assert.That(actual, Is.EqualTo(result));
         }
