+++ Nemesis.TextParsers/02_EntryPoint.cs
@@ -23,6 +23,7 @@ namespace Nemesis.TextParsers
 
     public static class TextTransformer
     {
+        //TODO provide mock object for tests + mock object with different settings 
         public static ITransformerStore Default { get; } =
             StandardTransformerStore.GetDefault(SettingsStoreBuilder.GetDefault().Build());
