+++ Nemesis.TextParsers/Parsers/09_Collection.cs
@@ -77,8 +77,85 @@ namespace Nemesis.TextParsers.Parsers
             => _kind = kind;
 
 
-        protected override TCollection ParseCore(in ReadOnlySpan<char> input) =>
-            (TCollection)SpanCollectionSerializer.DefaultInstance.ParseCollection<TElement>(input, _kind);
+        protected override TCollection ParseCore(in ReadOnlySpan<char> input)
+        {
+            var stream = ParseStream(input);
+            var capacity = Settings.DefaultCapacity;
+            
+            switch (_kind)
+            {
+                case CollectionKind.List:
+                case CollectionKind.ReadOnlyCollection:
+                    {
+                        var result = new List<TElement>(capacity);
+
+                        foreach (var part in stream)
+                            result.Add(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(
+                                 _kind == CollectionKind.List
+                                     ? (IReadOnlyCollection<TElement>)result
+                                     : result.AsReadOnly()
+                        );
+                    }
+                case CollectionKind.HashSet:
+                case CollectionKind.SortedSet:
+                    {
+                        ISet<TElement> result = _kind == CollectionKind.HashSet
+                            ? (ISet<TElement>)new HashSet<TElement>(
+#if NETSTANDARD2_0 || NETFRAMEWORK
+
+#else
+                            capacity
+#endif
+                        )
+                            : new SortedSet<TElement>();
+
+                        foreach (var part in stream)
+                            result.Add(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(IReadOnlyCollection<TElement>)result;
+                    }
+                case CollectionKind.LinkedList:
+                    {
+                        var result = new LinkedList<TElement>();
+
+                        foreach (var part in stream)
+                            result.AddLast(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(IReadOnlyCollection<TElement>)result;
+                    }
+                case CollectionKind.Stack:
+                    {
+                        var result = new Stack<TElement>(capacity);
+
+                        foreach (var part in stream)
+                            result.Push(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(IReadOnlyCollection<TElement>)result;
+                    }
+                case CollectionKind.Queue:
+                    {
+                        var result = new Queue<TElement>(capacity);
+
+                        foreach (var part in stream)
+                            result.Enqueue(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(IReadOnlyCollection<TElement>)result;
+                    }
+                case CollectionKind.ObservableCollection:
+                    {
+                        var result = new ObservableCollection<TElement>();
+
+                        foreach (var part in stream)
+                            result.Add(part.ParseWith(ElementTransformer));
+
+                        return (TCollection)(IReadOnlyCollection<TElement>)result;
+                    }
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(_kind), _kind, $"{nameof(_kind)} = '{nameof(CollectionKind)}.{nameof(CollectionKind.Unknown)}' is not supported");
+            }
+        }
 
 
         public override TCollection GetEmpty() =>
